# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: API specification for campanha politica CRM
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /regioes:
    get:
      operationId: listRegioes
      tags: [regioes]
      responses:
        "200":
          description: List of regions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RegiaoComStats"
    post:
      operationId: createRegiao
      tags: [regioes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRegiaoBody"
      responses:
        "201":
          description: Region created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Regiao"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /regioes/{id}:
    get:
      operationId: getRegiao
      tags: [regioes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Region
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegiaoDetalhe"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateRegiao
      tags: [regioes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRegiaoBody"
      responses:
        "200":
          description: Updated region
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Regiao"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRegiao
      tags: [regioes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /regioes/{id}/observacoes:
    get:
      operationId: listObservacoes
      tags: [observacoes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of observations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ObservacaoRegiao"
    post:
      operationId: createObservacao
      tags: [observacoes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateObservacaoBody"
      responses:
        "201":
          description: Observation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObservacaoRegiao"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Regiao:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        descricao:
          type: ["string", "null"]
        coordenador_regional_id:
          type: ["integer", "null"]
        cor:
          type: string
        prioridade:
          type: string
          enum: [normal, atencao, prioritaria]
        observacao_estrategica:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, nome, cor, prioridade, created_at, updated_at]
    RegiaoComStats:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        descricao:
          type: ["string", "null"]
        coordenador_regional_id:
          type: ["integer", "null"]
        coordenador_nome:
          type: ["string", "null"]
        cor:
          type: string
        prioridade:
          type: string
          enum: [normal, atencao, prioritaria]
        observacao_estrategica:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        total_contatos:
          type: integer
        total_simpatizantes:
          type: integer
        total_fechados:
          type: integer
        total_lideres:
          type: integer
      required: [id, nome, cor, prioridade, created_at, updated_at, total_contatos, total_simpatizantes, total_fechados, total_lideres]
    RegiaoDetalhe:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        descricao:
          type: ["string", "null"]
        coordenador_regional_id:
          type: ["integer", "null"]
        coordenador_nome:
          type: ["string", "null"]
        cor:
          type: string
        prioridade:
          type: string
          enum: [normal, atencao, prioritaria]
        observacao_estrategica:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        total_contatos:
          type: integer
        total_simpatizantes:
          type: integer
        total_fechados:
          type: integer
        total_lideres:
          type: integer
        lideres:
          type: array
          items:
            $ref: "#/components/schemas/StatsLider"
        observacoes:
          type: array
          items:
            $ref: "#/components/schemas/ObservacaoRegiao"
        proximos_eventos:
          type: array
          items:
            $ref: "#/components/schemas/Evento"
      required: [id, nome, cor, prioridade, created_at, updated_at, total_contatos, total_simpatizantes, total_fechados, total_lideres, lideres, observacoes, proximos_eventos]
    CreateRegiaoBody:
      type: object
      properties:
        nome:
          type: string
        descricao:
          type: ["string", "null"]
        coordenador_regional_id:
          type: ["integer", "null"]
        cor:
          type: ["string", "null"]
        prioridade:
          type: ["string", "null"]
        observacao_estrategica:
          type: ["string", "null"]
      required: [nome]
    UpdateRegiaoBody:
      type: object
      properties:
        nome:
          type: ["string", "null"]
        descricao:
          type: ["string", "null"]
        coordenador_regional_id:
          type: ["integer", "null"]
        cor:
          type: ["string", "null"]
        prioridade:
          type: ["string", "null"]
        observacao_estrategica:
          type: ["string", "null"]
    ObservacaoRegiao:
      type: object
      properties:
        id:
          type: integer
        regiao_id:
          type: integer
        autor_id:
          type: integer
        autor_nome:
          type: ["string", "null"]
        observacao:
          type: string
        created_at:
          type: string
          format: date-time
      required: [id, regiao_id, autor_id, observacao, created_at]
    CreateObservacaoBody:
      type: object
      properties:
        observacao:
          type: string
      required: [observacao]
    Evento:
      type: object
      properties:
        id:
          type: integer
        titulo:
          type: string
        descricao:
          type: ["string", "null"]
        data:
          type: string
        hora:
          type: ["string", "null"]
        local:
          type: ["string", "null"]
        tipo_evento:
          type: string
          enum: [reuniao, caminhada, visita, comicio, acao_de_rua, evento_interno]
        regiao_id:
          type: ["integer", "null"]
        regiao_nome:
          type: ["string", "null"]
        criado_por:
          type: integer
        criado_por_nome:
          type: ["string", "null"]
        visibilidade:
          type: string
          enum: [geral, regional, lider]
        coordenador_regional_id:
          type: ["integer", "null"]
        lider_id:
          type: ["integer", "null"]
        created_at:
          type: string
          format: date-time
      required: [id, titulo, data, tipo_evento, criado_por, visibilidade, created_at]
...
    StatsLider:
      type: object
      properties:
        lider_id:
          type: ["integer", "null"]
        lider_nome:
          type: ["string", "null"]
        total:
          type: integer
        simpatizantes:
          type: integer
        fechados:
          type: integer
      required: [total, simpatizantes, fechados]
...