# lib/api-zod/src/generated/types/createObservacaoBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for campanha politica CRM
 * OpenAPI spec version: 0.2.0
 */

export interface CreateObservacaoBody {
  observacao: string;
}