# lib/api-zod/src/generated/types/regiaoDetalhe.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for campanha politica CRM
 * OpenAPI spec version: 0.2.0
 */
import type { Evento } from "./evento";
import type { ObservacaoRegiao } from "./observacaoRegiao";
import type { RegiaoDetalhePrioridade } from "./regiaoDetalhePrioridade";
import type { StatsLider } from "./statsLider";

export interface RegiaoDetalhe {
  id: number;
  nome: string;
  /** @nullable */
  descricao?: string | null;
  /** @nullable */
  coordenador_regional_id?: number | null;
  /** @nullable */
  coordenador_nome?: string | null;
  cor: string;
  prioridade: RegiaoDetalhePrioridade;
  /** @nullable */
  observacao_estrategica?: string | null;
  created_at: Date;
  updated_at: Date;
  total_contatos: number;
  total_simpatizantes: number;
  total_fechados: number;
  total_lideres: number;
  lideres: StatsLider[];
  observacoes: ObservacaoRegiao[];
  proximos_eventos: Evento[];
}